feat: 지원가능 목록에도 필터링 적용
